requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

alembic==1.7.1; python_version >= '3.6'

atomicwrites==1.4.0; sys_platform == 'win32'

attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

authlib==0.15.4

bcrypt==3.2.0; python_version >= '3.6'

blinker==1.4

certifi==2021.5.30

cffi==1.14.6

charset-normalizer==2.0.4; python_version >= '3'

click==8.0.1; python_version >= '3.6'

colorama==0.4.4; sys_platform == 'win32' and platform_system == 'Windows'

cryptography==3.4.8; python_version >= '3.6'

dnspython==2.1.0; python_version >= '3.6'

ecdsa==0.17.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

email-validator==1.1.3

flask-bcrypt==0.7.1

flask-login==0.5.0

flask-mail==0.9.1

flask-migrate==3.1.0

flask-script==2.0.6

flask-sqlalchemy==2.5.1

flask-testing==0.8.1

flask-wtf==0.15.1

flask==2.0

greenlet==1.1.1; python_version >= '3' and platform_machine in 'x86_64 X86_64 aarch64 AARCH64 ppc64le PPC64LE amd64 AMD64 win32 WIN32'

gunicorn==20.1.0

idna==3.2; python_version >= '3'

importlib-resources==5.2.2; python_version < '3.9'

iniconfig==1.1.1

itsdangerous==2.0.1; python_version >= '3.6'

jinja2==3.0.1; python_version >= '3.6'

mako==1.1.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

markupsafe==2.0.1; python_version >= '3.6'

packaging==21.0; python_version >= '3.6'

pluggy==1.0.0; python_version >= '3.6'

psycopg2-binary==2.9.1

psycopg2==2.9.1

py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pyasn1==0.4.8

pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

pytest==6.2.5

python-dotenv==0.19.0

python-jose==3.3.0

requests==2.26.0

rsa==4.7.2; python_version >= '3.5' and python_version < '4'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlalchemy==1.4.23; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

werkzeug==2.0.1; python_version >= '3.6'

wtforms==2.3.3

zipp==3.5.0; python_version < '3.10'